third_party/blink/web_tests/fast/js/vardecl-preserve-arguments-expected.txt

This test checks that variable declaration 'var arguments;' doesn't override function's local arguments object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'object' is 'object'
PASS argumentsLength() is 0
PASS 'object' is 'object'
PASS argumentsLength(1) is 1
PASS 'object' is 'object'
PASS argumentsLength('a','b') is 2
PASS 'object' is 'object'
PASS argumentsLengthInnerBlock() is 0
PASS 'object' is 'object'
PASS argumentsLengthInnerBlock(1) is 1
PASS 'object' is 'object'
PASS argumentsLengthInnerBlock('a','b') is 2
PASS argumentsLengthInnerBlock2() is 0
PASS argumentsLengthInnerBlock2(1) is 1
PASS argumentsLengthInnerBlock2('a','b') is 2
PASS argumentsLengthTryCatch() is 0
PASS argumentsLengthWith() is 0
PASS 'object' is 'object'
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverride() is 0
PASS 'object' is 'object'
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverride(1) is 1
PASS 'object' is 'object'
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverride('a','b') is 2
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock() is 0
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock(1) is 1
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock('a','b') is 2
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock2() is 0
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock2(1) is 1
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock2('a','b') is 2
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock3() is 0
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock3(1) is 1
PASS 'object' is 'object'
PASS 7 is 7
PASS argumentsLengthOverrideInnerBlock3('a','b') is 2
PASS argumentsTearOff1() is 2
PASS typeof undefined is 'undefined'
PASS 'undefined' is 'undefined'
PASS typeof arguments is 'object'
PASS 'undefined' is 'undefined'
PASS 'object' is 'object'
PASS 2 is 2
PASS successfullyParsed is true

TEST COMPLETE
